LMDB: Fix param in add_block

+++ src/blockchain_db/lmdb/db_lmdb.h
@@ -367,8 +367,8 @@ private:
   void check_and_resize_for_batch(uint64_t batch_num_blocks, uint64_t batch_bytes);
   uint64_t get_estimated_batch_size(uint64_t batch_num_blocks, uint64_t batch_bytes) const;
 
-  virtual void add_block( const block& blk
-                , uint64_t block_weight
+ virtual void add_block( const block& blk
+                , size_t block_weight
                 , uint64_t long_term_block_weight
                 , const difficulty_type& cumulative_difficulty
                 , const uint64_t& coins_generated
